Move Layout to component folder
