+++ src/templates/general-post.js
@@ -1,5 +1,5 @@
 import React from "react"
-import Layout from "../components/layout"
+import Layout from "../components/Layout/Layout"
 import {graphql} from "gatsby"
 
 export default ({data, pageContext}) => (
